Record interrupt info read from shared area
This is an mechanism optimization, when interrupt
is triggered when the vcpu is not the lock owner,
the interrupt will be recorded by the guest recorder.
However, it needs to get the hardware info about the
interrupt like instruction count. Previously, this
info was stored in hypervisor and merged into the
event trace at the end of record session, now we
change the mechanism: there is a shared memory area
for hypervisor to write such info, and when guest
recorder records the interrupt, it reads from the shared
area, and the full info of the interrupt is written
in the event trace at runtime.